+++ ui/components/modals/RELEASENOTES.md
@@ -11,6 +11,10 @@
 - Added `slds` as the default namespace with `sds` fallbacks for Styling Hooks.
 - Added CSS Custom Property reassignments for variants and states.
 
+### Changed
+
+- Changed `header` and `footer` to generic `div` tags for better accessibility semantics
+
 ## 2.14.0
 
 ### Added